Clock: fixes 64 bit warnings

+++ src/apps/clock/cl_view.cpp
@@ -333,7 +333,7 @@ void
 TOnscreenView::MouseDown( BPoint point )
 {
 	BPoint	cursor;
-	ulong	buttons;
+	uint32	buttons;
 	BRect	bounds = Bounds();
 	
 	GetMouse(&cursor,&buttons);
